requirements.txt
%%writefile requirements.txt
torch
torchvision
streamlit
pillow
scikit-learn
